lntest: MOVEONLY networktest.go -> lntest package.
